pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.9",
    "chromadb>=1.0.16",
    "fastapi>=0.116.1",
    "fastembed>=0.7.1",
    "gradio>=5.42.0",
    "numpy>=2.3.2",
    "openai>=1.99.3",
    "packaging>=25.0",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.0",
    "uvicorn>=0.35.0",
]
